Move stores to popup folder
