+++ src/popup/stores/AppStore.ts
