+++ src/popup/stores/ImportStore.ts
@@ -2,7 +2,7 @@ import { observable, action, computed, reaction } from 'mobx';
 import { isEmpty } from 'lodash';
 
 import AppStore from './AppStore';
-import { MESSAGE_TYPE } from '../constants';
+import { MESSAGE_TYPE } from '../../constants';
 
 const INIT_VALUES = {
   mnemonic: '',
