+++ src/popup/stores/SessionStore.ts
@@ -1,7 +1,7 @@
 import { observable, action } from 'mobx';
 import { Insight } from 'qtumjs-wallet';
 
-import { MESSAGE_TYPE } from '../constants';
+import { MESSAGE_TYPE } from '../../constants';
 
 const INIT_VALUES = {
   networkIndex: 1,
